refactor(copilotkit): restructure server setup and update development script
